made the navigation bar on the index page and blog page use the same component
