+++ src/components/ProfileNavigation.jsx
@@ -65,4 +65,5 @@ export default function ProfileNavigation() {
       </>
     );
   }
+  return <></>;
 }
